+++ scss/linodes/linode/index.scss
@@ -1,2 +1,1 @@
 @import 'layouts/index.scss';
-@import 'settings/index.scss';
