--- scss/linodes/linode/settings/components/ConfigPanel.scss
@@ -1,3 +1,0 @@
-.ConfigPanel-configs tbody tr button {
-    margin-bottom: 3px;
-}
