--- scss/linodes/linode/settings/components/DiskPanel.scss
@@ -1,6 +1,0 @@
-.LinodesLinodeSettingsComponentsDiskPanel-edit {
-    margin-right: 5px;
-}
-.LinodesLinodeSettingsComponentsAddModal-disk-size {
-    width: 75px;
-}
