--- scss/linodes/linode/settings/components/index.scss
@@ -1,2 +1,0 @@
-@import 'ConfigPanel.scss';
-@import 'DiskPanel.scss';
