--- scss/linodes/linode/settings/disks.scss
@@ -1,32 +1,0 @@
-.disk-layout {
-    display: flex;
-    flex-wrap: wrap;
-    margin: 10px auto -5px auto;
-
-    .disk {
-        margin: 0 10px 10px 0;
-        padding: 10px;
-        background: $white;
-        border-size: 2px;
-        border-style: solid;
-        border-color: $border-gray;
-
-        &.free, &.disk-deleting {
-            background: transparent;
-            border: 5px dashed $border-gray !important;
-        }
-
-        h3 {
-            color: $light-black;
-        }
-
-        p {
-            color: $light-black;
-            margin: 0;
-        }
-
-        button {
-            margin-top: 10px;
-        }
-    }
-}
