--- scss/linodes/linode/settings/index.scss
@@ -1,17 +1,0 @@
-@import 'disks';
-@import 'components/index.scss';
-
-.measure-unit {
-    display: inline-block;
-    line-height: $form-field-height;
-    padding-left: 5px;
-}
-
-.multi-radio {
-    display: inline-block;
-    line-height: $form-field-height;
-
-    span {
-        padding-right: 10px;
-    }
-}
